+++ R/LearnerClustCMeans.R
@@ -82,14 +82,15 @@ LearnerClustCMeans = R6Class("LearnerClustCMeans",
 
   private = list(
     .train = function(task) {
-      if (test_data_frame(self$param_set$values$centers)) {
-        if (length(self$param_set$values$centers) != task$ncol) {
-          stop("`centers` must have same number of columns as data.")
-        }
-      }
+      check_centers_param(self$param_set$values$centers, task, test_data_frame, "centers")
 
       pv = self$param_set$get_values(tags = "train")
-      invoke(e1071::cmeans, x = task$data(), .args = pv, .opts = allow_partial_matching)
+      m = invoke(e1071::cmeans, x = task$data(), .args = pv, .opts = allow_partial_matching)
+      if (self$save_assignments) {
+        self$assignments = m$cluster
+      }
+
+      return(m)
     },
 
     .predict = function(task) {