+++ R/LearnerClustKMeans.R
@@ -65,14 +65,15 @@ LearnerClustKMeans = R6Class("LearnerClustKMeans",
         }
       }
 
-      if (test_data_frame(self$param_set$values$centers)) {
-        if (length(self$param_set$values$centers) != task$ncol) {
-          stop("`centers` must have same number of columns as data.")
-        }
-      }
+      check_centers_param(self$param_set$values$centers, task, test_data_frame, "centers")
 
       pv = self$param_set$get_values(tags = "train")
-      invoke(stats::kmeans, x = task$data(), .args = pv)
+      m = invoke(stats::kmeans, x = task$data(), .args = pv)
+      if (self$save_assignments) {
+        self$assignments = m$cluster
+      }
+
+      return(m)
     },
     .predict = function(task) {
       partition = unclass(cl_predict(self$model, newdata = task$data(), type = "class_ids"))