+++ NEWS.md
@@ -1,2 +1,6 @@
+## mlr3cluster 0.1.1
+*	Eight new learners
+*	Added `assignments` and `save_assignments` fields to `LearnerClust` class
+  
 ## mlr3cluster 0.1.0
-	- Initial upload to CRAN
+*	Initial upload to CRAN